refactor: started refactoring FormGenerator
